==> LBW Bayesian Modelling Code/0.fileNames.R <==
wpp2<-"WPP2022_Live_births_by_sex.xlsx"
birth_healthfacility<- "birth_healthfacility_who2021.dta"
adminName<-"input/Admin_Database_04March2022.xlsx"
newAdminName<-"input/LBW admin data for modelling 13 Dec 2022 - updated.xlsx"

surveyName<-"input/2022 12 01 LBW estimates combined database.xlsx"

regionsName<- "sbr_regions v2.csv"
