==> LBW Bayesian Modelling Code/0.loadPackages.R <==
library(boot)
library(DescTools)
library(devtools)
devtools::install_github("MJAlexander/distortr")
library(distortr)
library(dplyr)
library(fields)
library(foreign)
library(gridGraphics)
library(gtools)
library(haven)
library(meta)
library(RColorBrewer)
library(readxl)
library(scales)
library(splines)
library(tidyverse)
library(writexl)
#Ensure you have JAGS installed on your machine: https://sourceforge.net/projects/mcmc-jags/files/
library(rjags)
library(R2jags)
